+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for running Unity inside a container starting from 2023.09 TeamCity version
 - Support for Unity Personal license activation via .ulf file upload
+- Add option to remove '-quit' argument
 
 ## 1.0.4 - 2023-05-03